requirements.txt
# requirements.txt
# ─────────────────────────────────────────────────────────────
# Port Operations Analytics Warehouse — Python dependencies
#
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────

# ── Data generation (generate_data.py) ───────────────────────
numpy==1.26.4
pandas==2.2.2

# ── Database driver (load_data.py + FastAPI) ──────────────────
psycopg2-binary==2.9.9
asyncpg==0.29.0              # async Postgres driver for FastAPI

# ── FastAPI service (api/main.py) ────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0    # ASGI server
pydantic==2.7.1              # request / response models
python-dotenv==1.0.1         # .env loading
httpx==0.27.0                # for testing with TestClient
